refactor: PerformanceImage를 독립 domain context로 분리
PerformanceImage domain, repository contract, error code를 performance 하위에서 performanceimage context로 옮겨 기존 infra.persistence.performanceimage 경계와 이름을 맞춘다.

Constraint: DB schema, JPA entity package, API response shape, repository method signatures remain unchanged
Rejected: Keeping PerformanceImage under performance domain | it kept a detail-image context hidden under the performance aggregate package while infra already had a separate performanceimage adapter
Confidence: high
Scope-risk: moderate
Directive: PerformanceImage imports should use com.beat.domain.performanceimage.*; do not recreate performance.domain.PerformanceImage
Tested: ./gradlew clean compileJava compileKotlin test transitionBoundaryTest check --no-daemon
Not-tested: external consumers outside this repository that may import the old domain package
