+++ apis/src/main/java/com/beat/apis/performance/application/PerformanceService.java
@@ -1,7 +1,6 @@
 package com.beat.apis.performance.application;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
@@ -24,21 +23,21 @@ import com.beat.apis.performance.application.dto.performanceDetail.PerformanceDe
 import com.beat.apis.performance.application.dto.performanceDetail.PerformanceDetailResponse;
 import com.beat.apis.performance.application.dto.performanceDetail.PerformanceDetailScheduleResponse;
 import com.beat.apis.performance.application.dto.performanceDetail.PerformanceDetailStaffResponse;
-import com.beat.domain.schedule.service.ScheduleDomainService;
-import com.beat.domain.booking.repository.BookingRepository;
 import com.beat.domain.booking.domain.BookingStatus;
+import com.beat.domain.booking.repository.BookingRepository;
 import com.beat.domain.cast.domain.Cast;
 import com.beat.domain.cast.repository.CastRepository;
-import com.beat.domain.member.repository.MemberRepository;
 import com.beat.domain.member.domain.Member;
 import com.beat.domain.member.exception.MemberErrorCode;
-import com.beat.domain.performance.repository.PerformanceRepository;
+import com.beat.domain.member.repository.MemberRepository;
 import com.beat.domain.performance.domain.Performance;
-import com.beat.domain.performance.domain.PerformanceImage;
 import com.beat.domain.performance.exception.PerformanceErrorCode;
-import com.beat.domain.performance.repository.PerformanceImageRepository;
-import com.beat.domain.schedule.repository.ScheduleRepository;
+import com.beat.domain.performance.repository.PerformanceRepository;
+import com.beat.domain.performanceimage.domain.PerformanceImage;
+import com.beat.domain.performanceimage.repository.PerformanceImageRepository;
 import com.beat.domain.schedule.domain.Schedule;
+import com.beat.domain.schedule.repository.ScheduleRepository;
+import com.beat.domain.schedule.service.ScheduleDomainService;
 import com.beat.domain.staff.domain.Staff;
 import com.beat.domain.staff.repository.StaffRepository;
 import com.beat.domain.user.exception.UserErrorCode;
